day11 part2 ⭐

+++ src/main/java/com/hallietheswan/day11/Operation.java
@@ -1,5 +1,5 @@
 package com.hallietheswan.day11;
 
 public interface Operation {
-    int updateWorryLevel(int worryLevel);
+    long updateWorryLevel(long worryLevel);
 }
